refactor(shop): 상품목록 컴포넌트 섹션 기반 구조로 개선
