+++ apps/shop/src/features/product/components/ProductSectionList.module.scss
