+++ apps/shop/src/features/product/components/index.ts
@@ -1,10 +1,10 @@
 export { default as DetailContent } from './DetailContent';
-export { default as ProductList } from './ProductList';
 export { default as ProductCardBtn } from './ProductCardBtn';
 export { default as Option } from './Option';
-export { ProductListLoading } from './ProductListLoading';
+export { default as ProductImage } from './ProductImage';
 export { default as Description } from './Description';
 export { default as OptionList } from './OptionList';
-export { ProductSection } from './ProductSection';
-export { BannerSlide } from './BannerSlide';
-export { default as ProductImage } from './ProductImage';
+export * from './ProductListLoading';
+export * from './ProductSection';
+export * from './BannerSlide';
+export * from './ProductSectionList';
