+++ apps/shop/src/features/product/constants.ts
@@ -1,0 +1,5 @@
+export const SECTION_TITLE = {
+  VANS: '언제 어디서나 함께, 반스 시리즈',
+  NIKE: 'Just Do It, Nike 시리즈',
+  ALL: '모든 신발 둘러보기',
+} as const;
