+++ apps/shop/src/features/product/index.ts
@@ -2,3 +2,4 @@ export * from './components';
 export * from './hooks';
 export * from './types';
 export * from './apis';
+export * from './constants';
